Preliminary latex formatter
Very similar to qqxml, but a few functions are added, in particular,
handle works in the same way as in qqhtml, allowing for different
formatting of different tags.

I use a list of special tags ‘handlers_available’ that have separate
handlers. Otherwise, I use a default handler ‘handle_dummy’ which is
almost the exact copy of ‘handle’ in qqxml